changed PC to short

+++ src/PC.java
@@ -1,15 +1,15 @@
 public class PC {
-    private byte instToBeExec;
+    private short instToBeExec;
 
     public PC(){
         this.instToBeExec = 0;
     }
 
-    public byte getInstToBeExec() {
+    public short getInstToBeExec() {
         return instToBeExec;
     }
 
-    public void setInstToBeExec(byte instToBeExec) {
+    public void setInstToBeExec(short instToBeExec) {
         this.instToBeExec = instToBeExec;
     }
